HugChat/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
hugchat==0.4.19
idna==3.7
python-decouple==3.8
requests==2.32.3
requests-toolbelt==1.0.0
urllib3==2.2.2
